Version ready for steam develop

+++ SteamVersion/Assets/VoiceChat/Scripts/TestNetworkIdentity.cs
@@ -2,6 +2,6 @@
 using VOCASY;
 public class TestNetworkIdentity : MonoBehaviour, INetworkIdentity
 {
-    public uint NetworkId { get; set; }
+    public ulong NetworkId { get; set; }
     public bool IsLocalPlayer { get; set; }
 }
